requirements.txt
pandas
seaborn
matplotlib
pillow
numpy
tensorflow